pyproject.toml
[project]
name = "goit-de-hw-06"
version = "0.1.0"
description = ""
authors = [
    {name = "Sergiy Pavlyshyn",email = "pavlyshynserjiy@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "kafka-python (>=2.3.1,<3.0.0)",
    "pyspark (==3.5.1)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
